+++ packages/matrix/tests/UpSet.test.tsx
@@ -1,6 +1,5 @@
 import { render } from '@testing-library/react'
-import { cloneDeep } from 'lodash'
-import { Chart } from '@chsk/core'
+import { cloneProps, Chart } from '@chsk/core'
 import { UpSet, isUpSetProcessedData } from '../src/upset'
 import { upSetProps } from './props'
 import { GetProcessedData, mockProcessedData } from './contexts'
@@ -9,7 +8,7 @@ const ids = ['alpha', 'beta', 'gamma', 'delta']
 
 describe('UpSet', () => {
     it('defines processed data', () => {
-        const result = cloneDeep(mockProcessedData)
+        const result = cloneProps(mockProcessedData)
         render(
             <Chart>
                 <UpSet {...upSetProps}>