+++ packages/themes/src/complete/emptyTheme.ts
@@ -1,8 +1,7 @@
-import { CompleteThemeSpec, emptyTheme as coreEmptyTheme } from '@chsk/core'
-import { cloneDeep } from 'lodash'
+import { cloneProps, CompleteThemeSpec, emptyTheme as coreEmptyTheme } from '@chsk/core'
 
 // This theme is exactly the same as 'emptyTheme' in @chsk/core.
 // It is provided here for consistency with other complete themes in @chsk/themes.
 // With this definition, it is possible to import the empty theme from @chsk/themes
 
-export const emptyTheme: CompleteThemeSpec = cloneDeep(coreEmptyTheme)
+export const emptyTheme: CompleteThemeSpec = cloneProps(coreEmptyTheme)